# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: RoomOps API
  description: |
    API de gestion des réservations de salles pour espace de coworking B2B.
    
    ## Fonctionnalités
    - Réservation de salles avec gestion de conflits (optimistic locking + contrainte d’exclusion)
    - Gestion des pannes d’équipements (blocage automatique des salles)
    - Authentification JWT (access + refresh tokens)
    - Contrôle d’accès basé sur les rôles (Super-Admin, Manager, Employé)
    - Conformité RGPD (export des données, droit à l’oubli)
    - Visualisation des plannings et disponibilités
    
    ## Sécurité
    Tous les endpoints (sauf `/auth/*`) nécessitent un token JWT dans l’en-tête `Authorization: Bearer <token>`.
  version: 1.0.0
  contact:
    name: Support RoomOps
    email: supportroomops@gmail.com
servers:
  - url: http://localhost:8080/api/v1
    description: Serveur local de développement
...
paths:
...
  /bookings:
    post:
      tags:
        - Bookings
      summary: Créer une nouvelle réservation
      description: |
        Crée une réservation pour une salle sur un créneau horaire donné.
        Vérifie automatiquement la disponibilité (chevauchement) et les conflits.
        - Si un conflit est détecté, retourne 409 avec le code `BOOKING_CONFLICT`.
        - La salle doit être active et ne contenir aucun équipement en panne.
      operationId: createBooking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
            examples:
              validBooking:
                summary: Réservation valide
                value:
                  roomId: 1
                  dateDebut: '2024-01-15T09:00:00'
                  dateFin: '2024-01-15T11:00:00'
                  motif: "Réunion d'équipe projet"
      responses:
        '201':
          description: Réservation créée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Données invalides (dates incohérentes, salle inexistante, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Non authentifié
        '403':
          description: Rôle insuffisant
        '409':
          description: Conflit (créneau déjà réservé ou salle indisponible)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Bookings
      summary: Lister les réservations (avec pagination et filtres)
      description: |
        Récupère la liste des réservations selon les droits de l'utilisateur :
        - Employé : uniquement ses propres réservations
        - Manager : réservations de son entreprise
        - Super-Admin : toutes les réservations
      operationId: listBookings
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: query
          schema:
            type: integer
            format: int64
          description: Filtrer par salle
        - name: dateDebut
          in: query
          schema:
            type: string
            format: date-time
          description: Créneaux à partir de cette date
        - name: dateFin
          in: query
          schema:
            type: string
            format: date-time
          description: Créneaux jusqu'à cette date
        - name: statut
          in: query
          schema:
            $ref: '#/components/schemas/BookingStatut'
          description: Filtrer par statut
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Page (index 0)
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Nombre d'éléments par page
      responses:
        '200':
          description: Liste paginée des réservations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingPageResponse'
        '401':
          description: Non authentifié
  /bookings/{id}:
    get:
      tags:
        - Bookings
      summary: Obtenir une réservation par ID
      operationId: getBookingById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Détail de la réservation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '404':
          description: Réservation non trouvée

    put:
      tags:
        - Bookings
      summary: Modifier une réservation
      description: |
        Permet de changer la salle, les horaires ou le motif.
        Utilise l’optimistic locking : le champ `version` doit être celui actuellement en base.
        Vérifie les nouveaux conflits éventuels.
      operationId: updateBooking
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BookingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingUpdateRequest'
      responses:
        '200':
          description: Réservation modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Données invalides
        '409':
          description: Conflit de version ou de créneau

    delete:
      tags:
        - Bookings
      summary: Annuler une réservation (soft delete)
      description: Passe le statut à `ANNULEE` et libère le créneau.
      operationId: cancelBooking
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '204':
          description: Réservation annulée
        '404':
          description: Réservation non trouvée
...
  /rooms:
    get:
      tags:
        - Rooms
      summary: Lister toutes les salles
      description: Retourne la liste des salles avec leur statut (active, indisponible pour panne).
      operationId: listRooms
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Liste des salles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomResponse'
  /rooms/{roomId}:
    get:
      tags:
        - Rooms
      summary: Détail d'une salle
      operationId: getRoomById
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Salle trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '404':
          description: Salle non trouvée
  /rooms/{roomId}/availability:
    get:
      tags:
        - Rooms
      summary: Vérifier la disponibilité d'une salle sur un créneau
      operationId: checkRoomAvailability
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: dateDebut
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: dateFin
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: equipmentIds
          in: query
          schema:
            type: array
            items:
              type: integer
              format: int64
          description: |
            Identifiants des équipements requis pour l'usage prévu. Si fourni, la disponibilité
            applique les mêmes règles que la création de réservation : un équipement en panne
            ne bloque la salle que s'il est demandé ici.
      responses:
        '200':
          description: Disponibilité vérifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
        '400':
          description: Un ou plusieurs equipmentIds n'appartiennent pas à cette salle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Salle non trouvée

  # ============================
  # ÉQUIPEMENTS
  # ============================
...
  /users/me:
    get:
      tags:
        - Users
      summary: Profil de l'utilisateur connecté
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profil utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /users/export:
    get:
      tags:
        - Users
      summary: Export des données personnelles (RGPD)
      description: Retourne toutes les données liées à l'utilisateur connecté dans un format structuré.
      operationId: exportUserData
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Données exportées
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDataExport'
  /users/anonymize:
    delete:
      tags:
        - Users
      summary: Droit à l'oubli (anonymisation)
      description: |
        Anonymise les données personnelles de l'utilisateur (nom, prénom, email remplacés
        par des valeurs anonymisées). Les réservations (`Booking`) sont conservées telles
        quelles pour préserver l'intégrité de l'historique d'occupation des salles, mais
        elles sont détachées de toute donnée personnelle identifiante.
      operationId: anonymizeUser
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Anonymisation effectuée

# ============================================================
# COMPOSANTS RÉUTILISABLES
# ============================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via `/auth/login`
  parameters:
    BookingId:
      name: id
      in: path
      required: true
      description: Identifiant de la réservation
      schema:
        type: integer
        format: int64
  schemas:
...
    BookingRequest:
      type: object
      required:
        - roomId
        - dateDebut
        - dateFin
      properties:
        roomId:
          type: integer
          format: int64
          description: Identifiant de la salle
          example: 1
        dateDebut:
          type: string
          format: date-time
          example: '2024-01-15T09:00:00'
        dateFin:
          type: string
          format: date-time
          example: '2024-01-15T11:00:00'
        motif:
          type: string
          maxLength: 500
          description: Motif (optionnel)
          example: "Réunion d'équipe projet"
        equipmentIds:
          type: array
          items:
            type: integer
            format: int64
          description: |
            Identifiants des équipements de la salle réellement nécessaires pour cette
            réservation. Liste vide ou absente = aucun équipement requis ; un équipement en
            panne dans la salle ne bloque alors plus la réservation s'il n'est pas demandé.
    BookingUpdateRequest:
      type: object
      required:
        - version
      properties:
        roomId:
          type: integer
          format: int64
        dateDebut:
          type: string
          format: date-time
        dateFin:
          type: string
          format: date-time
        motif:
          type: string
          maxLength: 500
        equipmentIds:
          type: array
          items:
            type: integer
            format: int64
          description: |
            Identifiants des équipements requis pour cette réservation ; si fourni, remplace
            intégralement la liste précédemment associée à la réservation. Absent = liste
            existante conservée telle quelle, y compris en cas de changement de salle : si un
            équipement précédemment associé n'appartient plus à la nouvelle salle, la requête
            est rejetée en 400 (même règle de validation qu'une liste explicitement fournie).
            Le client doit alors refournir `equipmentIds` pour la nouvelle salle.
        version:
          type: integer
          format: int64
          description: Version actuelle (optimistic locking)
          example: 1
    BookingResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        roomId:
          type: integer
          format: int64
        roomName:
          type: string
        userId:
          type: integer
          format: int64
        userName:
          type: string
        companyName:
          type: string
        dateDebut:
          type: string
          format: date-time
        dateFin:
          type: string
          format: date-time
        statut:
          $ref: '#/components/schemas/BookingStatut'
        motif:
          type: string
        equipmentsActives:
          type: array
          items:
            $ref: '#/components/schemas/EquipmentResponse'
          description: Équipements de la salle réellement demandés pour cette réservation.
        ecoScore:
          type: integer
          minimum: 0
          maximum: 100
          description: |
            Pourcentage des équipements de la salle NON activés pour cette réservation. Vaut
            100 par convention si la salle ne possède aucun équipement.
          example: 75
        version:
          type: integer
          format: int64
    BookingPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BookingResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer
    BookingStatut:
      type: string
      enum:
        - CONFIRMEE
        - ANNULEE
      description: |
        Statut de la réservation. Pas de workflow d'approbation dans le MVP :
        une réservation est confirmée immédiatement si aucun conflit n'est détecté.

    # ---- ROOM ----
    RoomResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nom:
          type: string
          example: "Salle Alpha"
        capacite:
          type: integer
          example: 8
        buildingId:
          type: integer
          format: int64
        buildingName:
          type: string
        estActif:
          type: boolean
        indisponibilite:
          type: string
          nullable: true
          description: "Raison si la salle est bloquée (ex: 'Panne projecteur')"
          example: null
        equipements:
          type: array
          items:
            $ref: '#/components/schemas/EquipmentResponse'
          description: |
            Équipements disponibles dans cette salle, pour permettre au frontend de proposer
            le choix éco-responsable.
    AvailabilityResponse:
      type: object
      properties:
        roomId:
          type: integer
          format: int64
        roomName:
          type: string
        isAvailable:
          type: boolean
          description: Disponible sur le créneau demandé
        reason:
          type: string
          nullable: true
          description: Motif d'indisponibilité (panne, réservé, etc.)
        dateDebut:
          type: string
          format: date-time
        dateFin:
          type: string
          format: date-time

    # ---- EQUIPMENT ----
    EquipmentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          example: "Projecteur"
        roomId:
          type: integer
          format: int64
        roomName:
          type: string
        statut:
          $ref: '#/components/schemas/EquipmentStatut'
    EquipmentStatut:
      type: string
      enum:
        - OPERATIONNEL
        - EN_PANNE

    # Réponse dédiée à PUT /equipments/{id}/status : distincte d'EquipmentResponse (utilisée
    # aussi par GET /equipments en liste) pour ne pas publier un reservationsAnnulees=0 ambigu
    # sur les éléments de la liste, où ce compte n'a pas de sens.
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        nom:
          type: string
        prenom:
          type: string
        role:
          type: string
        companyId:
          type: integer
          format: int64
          nullable: true
        companyName:
          type: string
          nullable: true

    # ---- RGPD ----
    UserDataExport:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        bookings:
          type: array
          items:
            $ref: '#/components/schemas/BookingResponse'
        exportDate:
          type: string
          format: date-time

    # ---- ERREUR ----
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
          example: "BOOKING_CONFLICT"
        message:
          type: string
          example: "Cette salle est déjà réservée sur le créneau demandé"
        timestamp:
          type: string
          format: date-time
          example: '2024-01-15T08:30:00'
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string